+++ Assets/Mirror/Tests/Editor/NetworkServerTest.cs
@@ -1077,5 +1077,12 @@ namespace Mirror.Tests
             Assert.That(NetworkServer.localConnection, Is.Null);
             Assert.That(NetworkServer.localClientActive, Is.False);
         }
+
+        [Test]
+        public void ResetTest()
+        {
+            NetworkServer.Reset();
+            Assert.That(NetworkServer.active, Is.False);
+        }
     }
 }