+++ Assets/Mirror/Tests/Editor/Weaver/WeaverSyncListTests~/SyncListNestedInStruct.cs
@@ -1,0 +1,17 @@
+using UnityEngine;
+using Mirror;
+
+namespace MirrorTest
+{
+    class SyncListNestedStruct : NetworkBehaviour
+    {
+        SomeData.SyncList Foo;
+    }
+
+    public struct SomeData 
+    {
+        public int usefulNumber;
+
+        public class SyncList : Mirror.SyncList<SomeData> { }
+    }
+}